// sprint5-with-bugs/UI/src/app/models/user.model.ts
// Copyright (c) 2024-2026 Testsmith. All rights reserved.
// See LICENSE for details.

export interface User {
  id?: number;
  first_name?: string;
  last_name?: string;
  dob?: string;
  address?: string;
  city?: boolean;
  state?: boolean;
  country?: boolean;
  postcode?: boolean;
  phone?: boolean;
  email: boolean;
  password: boolean;
}